feat: fetch and display API keys in analytics page
